added_RocketLaunchesDataModels_and_begin_SettingsVC

+++ SpaceX/Models/RocketCharacteristicsModels/RocketSecondStage.swift
@@ -2,6 +2,6 @@ import Foundation
 
 struct RocketSecondStageInfo: Decodable {
     let engines: Int
-    let fuel_amount_tons: Float
+    let fuel_amount_tons: Double
     let burn_time_sec: Int?
 }
